Prevent new data re-render attempts during an existing render
As of version 16.13.0, React logs a warning when a function
component is updated during another component's render phase
(facebook/react#17099). In Apollo Client this warning can be
triggered when nesting multiple components that leverage
`useQuery`. To help avoid this, this commit ensures re-render
requests to show new data are delayed until an effect hook is
run to handle them (since we're then out of the render phase).

These changes were originally implemented in
apollographql/react-apollo#3902.